Move database quote maintenance into autopilot
